% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDBoot.R
\name{SDBoot}
\alias{SDBoot}
\title{Subsampled Double Bootstrap}
\usage{
SDBoot(dta, statistic, T, subset_size, ..., time_lim = 300)
}
\arguments{
\item{dta}{The data you want to bootstrap.}

\item{statistic}{The statistic of interest (theta_hat).}

\item{T}{The "root" or error function of interest.
Note that the first argument should be the theta of the sample, and the senond arguent should be the
"true" theta.}

\item{subset_size}{How large should the initial subset be.}

\item{...}{Extra parameters to pass into T.star or FUN.}

\item{time_lim}{How long should the function run. Default is 5 minutes.}
}
\value{
List of the theta_hat of dta, a vector containing all the T.star
and the number of iterations completed in the time limit
}
\description{
Given the data, some function of the data, and some error function,
runs the Subsampled Double bootstrap algorithm.
}
